+++ example/main.cpp
@@ -3,5 +3,6 @@
 
 int main (int argc, char* argv [])
 {
+	setlocale (LC_ALL, "");
 	return IDL_Print ().main (argc, argv);
 }